Create Teensy Audio Test Code
This is test code used to verify the functionality of the Teensy 3.1, the Audio Adapter Board Rev B, and the SD card. All are working, and the program successfully reads and plays the wav files on the SD card. I also determined the Teensy can mix up to 3 notes at a time for simultaneous play.